Move ConfigQueueingPanel and Preference widgets into individual class files
